// admin/src/lib/types/api/UpdateCircleFormRequest.ts
export interface UpdateCircleFormRequest {
  type: 'UpdateCircleFormRequest'
  slug: string
  release: boolean
  isMainFixed?: boolean
  isOnlyDemo?: boolean
  isDemoFixed?: boolean
  demoPriority?: number
  name: string
  nameKana?: string
  shortName?: string
  prefixName?: string
  circleType?: string
  description?: string
  commonPlaceOfActivity?: string
  commonPlaceOfActivityDetail?: string
  commonDateOfActivityMonday?: boolean
  commonDateOfActivityTuesday?: boolean
  commonDateOfActivityWednesday?: boolean
  commonDateOfActivityThursday?: boolean
  commonDateOfActivityFriday?: boolean
  commonDateOfActivitySaturday?: boolean
  commonDateOfActivitySunday?: boolean
  onlineDateOfActivityDetail?: string
  isOnlineActivity?: boolean
  onlinePlaceOfActivityDetail?: string
  onlineDateOfActivityMonday?: boolean
  onlineDateOfActivityTuesday?: boolean
  onlineDateOfActivityWednesday?: boolean
  onlineDateOfActivityThursday?: boolean
  onlineDateOfActivityFriday?: boolean
  onlineDateOfActivitySaturday?: boolean
  onlineDateOfActivitySunday?: boolean
  commonDateOfActivityDetail?: string
  admissionFeePerYear?: number
  numberOfMembers?: number
  isClubActivities?: boolean
  appealingPoint1?: string
  appealingPoint2?: string
  appealingPoint3?: string
  publicEmail?: string
  twitterUrl?: string
  facebookUrl?: string
  instagramUrl?: string
  lineUrl?: string
  youtubeUrl?: string
  homepageUrl?: string
  peingUrl?: string
  githubUrl?: string
  tiktokUrl?: string
  participationUrl?: string
  mainImageUrl?: string
  activityImageUrl1?: string
  activityImageUrl2?: string
  activityImageUrl3?: string
  activityImageUrl4?: string
  activityImageUrl5?: string
  activityImageUrl6?: string
  wpUrl?: string
  wpTagTaxonomy?: string
  isViewWpPost?: boolean
  handbillImageUrl?: string
}